requirements.txt
asgiref==3.3.4
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-ckeditor==6.0.0
django-countries==7.1
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-phone-field==1.8.1
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
idna==2.10
phonenumbers==8.12.21
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
qrcode==6.1
requests==2.25.1
secure-smtplib==0.1.1
six==1.15.0
sqlparse==0.4.1
twilio==6.56.0
urllib3==1.26.4
